requirements.txt
fastapi==0.115.5
psycopg2-binary==2.9.10
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
uvicorn[standard]
langchain-ollama==0.2.0
passlib[bcrypt] 
python-jose==3.3.0
